test: Adding CastMemberController tests

+++ database/migrations/2021_05_06_000700_create_cast_members_table.php
@@ -16,7 +16,7 @@ class CreateCastMembersTable extends Migration
         Schema::create('cast_members', function (Blueprint $table) {
             $table->uuid('id')->primary();
             $table->string('name');
-            $table->bigInteger('type');
+            $table->smallInteger('type');
             $table->softDeletes();
             $table->timestamps();
         });
